refactor: :art: Refactor cronjobs code to organize in much better way
