+++ app/src/app/cron-jobs/page.tsx
@@ -1,36 +1,16 @@
 "use client";
-
 import React, { useState } from 'react';
-import { Button } from '@/components/ui/button';
-import { PlusIcon, Search } from 'lucide-react';
 import PaginateComp from '@/components/paginate';
-import { Input } from "@/components/ui/input";
-import CronJobCard from './components/CronJobCard';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { useSortableSearchableData } from './hooks/useSortableSearchData';
 import { useRouter } from 'next/navigation';
 import DashboardLayout from '../home/page';
 import CreateCronJob from './create/page';
 import { useGetAllCronJobsQuery } from '@/redux/services/cronJobsApi';
+import { CronJob } from './types';
+import CronJobList from './components/CronJobsList';
+import CronJobsHeader from './components/CronJobsHeader';
 
 const ITEMS_PER_PAGE = 8;
-
-export interface CronJob {
-    id: string;
-    name: string;
-    description: string;
-    schedule: string;
-    command: string;
-    is_active: boolean;
-    last_run: string;
-    next_run: string;
-    tag: string;
-}
-
-export enum JobStatus {
-    RUNNING = 'Running',
-    STOPPED = 'Stopped',
-}
 
 function CronJobs() {
     const [currentPage, setCurrentPage] = useState(1);
@@ -53,74 +33,41 @@ function CronJobs() {
         currentPage * ITEMS_PER_PAGE
     );
 
-    const paginatedJobs2: any[] = [];
-
-    const handlePageChange = (pageNumber: React.SetStateAction<number>) => {
+    const handlePageChange = (pageNumber: number) => {
         setCurrentPage(pageNumber);
     };
 
-    function onEditCronJob(id: string) {
+    const onEditCronJob = (id: string) => {
         router.push('/cron-jobs/' + id);
-    }
+    };
 
     if (isLoading) {
         return <div>Loading...</div>;
     }
 
-    return (
-        <>
-            {paginatedJobs2.length > 0 ? (
-                <DashboardLayout>
-                    <div className='flex flex-col '>
-                        <div className='flex justify-between items-center'>
-                            <div className='flex items-center my-10 space-x-4'>
-                                <h1 className='text-3xl font-bold'>Cron Jobs</h1>
-                                <div className="relative">
-                                    <Search className='absolute left-2 top-1/2 transform -translate-y-1/2 h-4 w-4 text-muted-foreground' />
-                                    <Input
-                                        type="text"
-                                        placeholder="Search jobs..."
-                                        value={searchTerm}
-                                        onChange={handleSearchChange}
-                                        className="pl-8 w-64"
-                                    />
-                                </div>
-                                <Select
-                                    onValueChange={(value) => handleSortChange(value.split('_')[0] as 'name' | 'tag')}
-                                    defaultValue="name_asc"
-                                >
-                                    <SelectTrigger className="w-[180px]">
-                                        <SelectValue placeholder="Sort by" />
-                                    </SelectTrigger>
-                                    <SelectContent>
-                                        <SelectItem value="name_asc">Name (A-Z)</SelectItem>
-                                        <SelectItem value="name_desc">Name (Z-A)</SelectItem>
-                                        <SelectItem value="tag_asc">Status (A-Z)</SelectItem>
-                                        <SelectItem value="tag_desc">Status (Z-A)</SelectItem>
-                                    </SelectContent>
-                                </Select>
-                            </div>
-                            <Button variant={'outline'} onClick={() => router.push('/cron-jobs/create')} >Create <PlusIcon className='ml-2 h-4 w-4' /></Button>
-                        </div>
-                        <div className='grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-8'>
-                            {paginatedJobs2.map((job, index) => (
-                                <CronJobCard key={index} job={job} JobStatus={JobStatus} onEditCronJob={onEditCronJob} />
-                            ))}
-                        </div>
-                        <div className="mt-8 flex justify-center">
-                            <PaginateComp
-                                currentPage={currentPage}
-                                totalPages={totalPages}
-                                onPageChange={handlePageChange}
-                            />
-                        </div>
-                    </div>
-                </DashboardLayout>
-            ) : (
-                <CreateCronJob />
-            )}
-        </>
+    if ((paginatedJobs || []).length === 0) {
+        return <CreateCronJob />;
+    }
 
+    return (
+        <DashboardLayout>
+            <div className='flex flex-col'>
+                <CronJobsHeader
+                    searchTerm={searchTerm}
+                    handleSearchChange={handleSearchChange}
+                    handleSortChange={handleSortChange}
+                    onCreateClick={() => router.push('/cron-jobs/create')}
+                />
+                <CronJobList jobs={paginatedJobs} onEditCronJob={onEditCronJob} />
+                <div className="mt-8 flex justify-center">
+                    <PaginateComp
+                        currentPage={currentPage}
+                        totalPages={totalPages}
+                        onPageChange={handlePageChange}
+                    />
+                </div>
+            </div>
+        </DashboardLayout>
     );
 }
